move dae to separate module
