+++ tests/devices/test_dae.py
@@ -1,7 +1,7 @@
 # pyright: reportMissingParameterType=false
 
 import pytest
-from ibex_bluesky_core.devices.dae import Dae, RunstateEnum
+from ibex_bluesky_core.devices.dae.dae import Dae, RunstateEnum
 from ophyd_async.core import get_mock_put
